128-bit precision on bigFloat

+++ ch03/ex08/c128/main.go
@@ -36,10 +36,10 @@ func mandelbrot(z complex128) color.Color {
 	const contrast = 15
 
 	var v complex128
-	for n := 0; n < iterations; n++ {
+	for n := uint8(0); n < iterations; n++ {
 		v = v*v + z
 		if cmplx.Abs(v) > 2 {
-			return color.Gray{255 - uint8(contrast*n)}
+			return color.Gray{255 - contrast*n}
 		}
 	}
 	return color.Black
